Refactor listener init arguments validation logic
